requirements.txt
fastapi==0.110.0
uvicorn==0.35.0
pillow==10.2.0